refactor: reuse openai image interface
